+++ CBash/FalloutNewVegas/Records/API/HAIRRecordAPI.cpp
@@ -88,7 +88,7 @@ UINT32 HAIRRecord::GetFieldAttribute(FIELD_IDENTIFIERS, UINT32 WhichAttribute)
                     case 0: //fieldType
                         return LIST_FIELD;
                     case 1: //fieldSize
-                        return MODL->Textures.MODS.size();
+                        return (UINT32)MODL->Textures.MODS.size();
                     default:
                         return UNKNOWN_FIELD;
                     }
