+++ CBash/FalloutNewVegas/Records/API/NOTERecordAPI.cpp
@@ -100,7 +100,7 @@ UINT32 NOTERecord::GetFieldAttribute(FIELD_IDENTIFIERS, UINT32 WhichAttribute)
                     case 0: //fieldType
                         return LIST_FIELD;
                     case 1: //fieldSize
-                        return MODL->Textures.MODS.size();
+                        return (UINT32)MODL->Textures.MODS.size();
                     default:
                         return UNKNOWN_FIELD;
                     }
@@ -139,7 +139,7 @@ UINT32 NOTERecord::GetFieldAttribute(FIELD_IDENTIFIERS, UINT32 WhichAttribute)
                 case 0: //fieldType
                     return FORMID_ARRAY_FIELD;
                 case 1: //fieldSize
-                    return ONAM.value.size();
+                    return (UINT32)ONAM.value.size();
                 default:
                     return UNKNOWN_FIELD;
                 }
